tests/wpt/tests/html/canvas/offscreen/fill-and-stroke-styles/2d.pattern.paint.orientation.image.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.pattern.paint.orientation.image</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.pattern.paint.orientation.image</h1>
<p class="desc">Image patterns do not get flipped when painted</p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);

  var response = await fetch('/images/rrgg-256x256.png')
  var blob = await response.blob();
  var img = await createImageBitmap(blob);
  var pattern = ctx.createPattern(img, 'no-repeat');
  ctx.fillStyle = pattern;
  ctx.save();
  ctx.translate(0, -103);
  ctx.fillRect(0, 103, 100, 50);
  ctx.restore();

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 25);

  _assertPixel(canvas, 1,1, 0,255,0,255);
  _assertPixel(canvas, 98,1, 0,255,0,255);
  _assertPixel(canvas, 1,48, 0,255,0,255);
  _assertPixel(canvas, 98,48, 0,255,0,255);

}, "Image patterns do not get flipped when painted");
</script>
